+++ __docs__/wonder-blocks-icon-button/icon-button-testing-snapshots.stories.tsx
@@ -8,7 +8,7 @@ import {AllVariants} from "../components/all-variants";
 import {defaultPseudoStates, StateSheet} from "../components/state-sheet";
 import {sizing} from "@khanacademy/wonder-blocks-tokens";
 import {View} from "@khanacademy/wonder-blocks-core";
-import {themeModes} from "../../.storybook/modes";
+import {allThemeModes} from "../../.storybook/modes";
 import {ScenariosLayout} from "../components/scenarios-layout";
 import {Icon} from "@khanacademy/wonder-blocks-icon";
 
@@ -25,11 +25,6 @@ export default {
         onClick: action("clicked"),
         actionType: "progressive",
         size: "medium",
-    },
-    parameters: {
-        chromatic: {
-            modes: themeModes,
-        },
     },
 } as Meta;
 
@@ -74,6 +69,9 @@ export const StateSheetStory: StoryComponentType = {
     },
     parameters: {
         pseudo: defaultPseudoStates,
+        chromatic: {
+            modes: allThemeModes,
+        },
     },
 };