+++ packages/wonder-blocks-tokens/src/theme/semantic/semantic-color-thunderblocks.ts
@@ -146,6 +146,10 @@ const core = {
                 default: color.gray_30,
                 strong: color.gray_10,
             },
+            disabled: {
+                subtle: transparent,
+                default: color.gray_60,
+            },
         },
     },
 };
@@ -525,9 +529,9 @@ export const semanticColor = mergeTheme(defaultSemanticColor, {
                 tertiary: color.gray_50,
             },
             shadow: {
-                primary: core.shadow.chonky.neutral.subtle,
-                secondary: core.shadow.chonky.neutral.subtle,
-                tertiary: core.transparent,
+                primary: core.shadow.chonky.disabled.default,
+                secondary: core.shadow.chonky.disabled.default,
+                tertiary: core.shadow.chonky.disabled.subtle,
             },
         },
     },
@@ -669,6 +673,35 @@ export const semanticColor = mergeTheme(defaultSemanticColor, {
                     default: color.green_60,
                 },
             },
+            characters: {
+                flesh: {
+                    blue: {
+                        subtle: color.blue_40,
+                        default: color.blue_30,
+                        strong: color.blue_20,
+                    },
+                    cyan: {
+                        subtle: color.cyan_40,
+                        default: color.cyan_30,
+                        strong: color.cyan_20,
+                    },
+                    green: {
+                        subtle: color.green_40,
+                        default: color.green_30,
+                        strong: color.green_20,
+                    },
+                    magenta: {
+                        subtle: color.magenta_40,
+                        default: color.magenta_30,
+                        strong: color.magenta_20,
+                    },
+                    orange: {
+                        subtle: color.orange_40,
+                        default: color.orange_30,
+                        strong: color.orange_20,
+                    },
+                },
+            },
         },
         border: {
             gems: {
@@ -685,6 +718,35 @@ export const semanticColor = mergeTheme(defaultSemanticColor, {
                 subtle: color.cyan_80,
                 default: color.cyan_60,
                 strong: color.cyan_30,
+            },
+            characters: {
+                flesh: {
+                    blue: {
+                        subtle: color.blue_40,
+                        default: color.blue_20,
+                        strong: color.blue_10,
+                    },
+                    cyan: {
+                        subtle: color.cyan_40,
+                        default: color.cyan_20,
+                        strong: color.cyan_10,
+                    },
+                    green: {
+                        subtle: color.green_40,
+                        default: color.green_20,
+                        strong: color.green_10,
+                    },
+                    magenta: {
+                        subtle: color.magenta_40,
+                        default: color.magenta_20,
+                        strong: color.magenta_10,
+                    },
+                    orange: {
+                        subtle: color.orange_40,
+                        default: color.orange_20,
+                        strong: color.orange_10,
+                    },
+                },
             },
         },
         foreground: {