+++ packages/wonder-blocks-modal/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @khanacademy/wonder-blocks-modal
 
+## 8.7.5
+
+### Patch Changes
+
+- Updated dependencies [f13552a]
+    - @khanacademy/wonder-blocks-tokens@16.8.0
+    - @khanacademy/wonder-blocks-breadcrumbs@3.2.22
+    - @khanacademy/wonder-blocks-icon-button@11.3.5
+    - @khanacademy/wonder-blocks-layout@3.1.55
+    - @khanacademy/wonder-blocks-styles@0.2.48
+    - @khanacademy/wonder-blocks-typography@4.3.8
+    - @khanacademy/wonder-blocks-announcer@1.1.1
+
+## 8.7.4
+
+### Patch Changes
+
+- c033245: Update modal and drawer panel background color to `semanticColor.core.background.base.default` to align with Figma specs
+- 29ae44b: Update modal and drawer panel background color to `semanticColor.core.background.base.subtle` to align with Figma specs
+- Updated dependencies [0907f64]
+- Updated dependencies [9773405]
+- Updated dependencies [9773405]
+- Updated dependencies [167298c]
+- Updated dependencies [022e775]
+- Updated dependencies [e181c15]
+- Updated dependencies [c033245]
+- Updated dependencies [a2bbc0a]
+    - @khanacademy/wonder-blocks-tokens@16.7.0
+    - @khanacademy/wonder-blocks-icon-button@11.3.4
+    - @khanacademy/wonder-blocks-breadcrumbs@3.2.21
+    - @khanacademy/wonder-blocks-layout@3.1.54
+    - @khanacademy/wonder-blocks-styles@0.2.47
+    - @khanacademy/wonder-blocks-typography@4.3.7
+    - @khanacademy/wonder-blocks-announcer@1.1.1
+
 ## 8.7.3
 
 ### Patch Changes