+++ packages/wonder-blocks-progress-spinner/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-progress-spinner
 
+## 3.1.55
+
+### Patch Changes
+
+- Updated dependencies [f13552a]
+    - @khanacademy/wonder-blocks-tokens@16.8.0
+
+## 3.1.54
+
+### Patch Changes
+
+- 79e872f: CircularSpinner: Use semantic color tokens for the spinner
+- Updated dependencies [0907f64]
+- Updated dependencies [9773405]
+- Updated dependencies [9773405]
+- Updated dependencies [022e775]
+- Updated dependencies [e181c15]
+- Updated dependencies [c033245]
+- Updated dependencies [a2bbc0a]
+    - @khanacademy/wonder-blocks-tokens@16.7.0
+
 ## 3.1.53
 
 ### Patch Changes